Workaround to solve a dependency problem in the deserialization of objects.
Currently the TomP2P classes need to be in the same classpath as the
client application in order to deserialize transferred objects
correctly. As a quick fix the Hive2Hive core includes the TomP2P jar
directly to overcome the problem. This workaround should be resolved as
quickly as possible to get a clean dependency structure.