Design and implementation of file events in response to remote actions such as uploading a file, deleting a file etc. The implementation is based on an EventBus rather than traditional Java observers/listeners because it decouples the listeners and generators of events much betters and most events happen deep down in the library..
The event bus instance is passed further down in order to be able to publish events. Arbitrary events can be published on the bus.
Events are then delivered to registered listeners and  handled by their annotated handler methods (@Handler) having the right parameters (i.e. the depending on the event type, different interfaces/implementations may be used and different methods will be called).

The bus supports synchronous as well as asynchronous event delivery (however, only synchronous events are currently used).

Library: MBassador is used as event bus, which is similar to the Guava event bus, but does not require the huge Guava library.